depgraph/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
websockets==13.1
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
tree-sitter-javascript==0.23.1
sqlglot==25.29.0
networkx==3.4.2
openai>=1.55.0
python-dotenv==1.0.1
gitpython==3.1.43

# Knowledge Graph / LangChain stack
langchain==0.3.25
langchain-openai==0.3.16
langchain-experimental==0.3.4
langchain-neo4j==0.4.0
langchain-core==0.3.65

# Neo4j driver
neo4j==5.27.0
